Updated Shared Libraries

+++ mymodules/operations/action.c
@@ -1,6 +1,6 @@
-#include "vectordot.c"
-float * action(float** matrix, float* vector, int s1, int s2) {
-    float* result = (float*)malloc(s1 * sizeof(float));
+#include <stdlib.h>
+double * action(double** matrix, double* vector, int s1, int s2) {
+    double* result = (double*)malloc(s1 * sizeof(double));
     for (int i = 0; i < s1; i++) {
         result[i] = vectorDot(matrix[i], vector, s1);
     }
