+++ mymodules/operations/daction.c
@@ -1,5 +1,7 @@
-float * dAction(float** matrix, float* vector, int s1, int s2) {
-    float* result = (float*)malloc(s1 * sizeof(float));
+#include <stdlib.h>
+
+double * dAction(double** matrix, double* vector, int s1, int s2) {
+    double* result = (double*)malloc(s1 * sizeof(double));
     for (int i = 0; i < s1; i++) {
         result[i] = matrix[i][i] * vector[i];
     }
