+++ mymodules/operations/jacoby.c
@@ -3,27 +3,24 @@
 #include <omp.h>
 #include <math.h>
 
-#include "../operations/vectoradd.c"
-#include "../operations/vectorsub.c"
-#include "../operations/daction.c"
 
-float * jacoby(float** A,  float* x0, float* b, double tol, int maxIter, int s1, double* time) {
+double * jacoby(double** A,  double* x0, double* b, double tol, int maxIter, int s1, double* time) {
     double startTime = omp_get_wtime();
 
 
     int iter = 0;
     double error = tol * 10;
-    float** D = (float**)malloc(s1 * sizeof(float*));
+    double** D = (double**)malloc(s1 * sizeof(double*));
     for (int i = 0; i < s1; i++) {
-        D[i] = (float*)malloc(s1 * sizeof(float));
+        D[i] = (double*)malloc(s1 * sizeof(double));
     }
 
     for (int i = 0; i < s1; i++) {
         D[i][i] = 1 / A[i][i];
     }
 
-    float* r0 = vectorSub(b, actionOMP(A, x0, s1, s1), s1);
-    float* x1;
+    double* r0 = vectorSub(b, actionOMP(A, x0, s1, s1), s1);
+    double* x1;
     while (iter < maxIter && error > tol) {
         error = norm(r0, s1);
         x1 = vectorAdd(x0, dAction(D, r0, s1, s1), s1);
