+++ mymodules/operations/norm.c
@@ -1,7 +1,8 @@
 #include <math.h>
+#include <stdlib.h>
 
-float norm(float* v, int size) {
-    float result = 0;
+double norm(double* v, int size) {
+    double result = 0;
     for (int i = 0; i < size; i++) {
         result += v[i] * v[i];
     }
