+++ mymodules/operations/vectoradd.c
@@ -1,5 +1,7 @@
-float * vectorAdd(float* v1, float* v2, int s1) {
-    float* result = (float*)malloc(s1 * sizeof(float));
+#include <stdlib.h>
+
+double * vectorAdd(double* v1, double* v2, int s1) {
+    double* result = (double*)malloc(s1 * sizeof(double));
     for (int i = 0; i < s1; i++) {
         result[i] =  v1[i] + v2[i];
     }
